--- Lab1/backup/sudoku.h
@@ -1,26 +1,0 @@
-#ifndef SUDOKU_H
-#define SUDOKU_H
-
-const bool DEBUG_MODE = false;
-//NEIGHBOR: for a certain digit, it needs to match with 20 ambient digits which are at the colomn(6), row(6) and board(8).
-enum { ROW=9, COL=9, N = 81, NEIGHBOR = 20 };
-const int NUM = 9;
-
-//extern int neighbors[N][NEIGHBOR];
-//extern int board[N];
-//extern int spaces[N];
-//extern int nspaces;
-//extern int (*chess)[COL];
-
-void init_neighbors();
-void input(const char in[N]);
-void init_cache();
-
-bool available(int guess, int cell);
-
-bool solve_sudoku_basic(int which_space);
-bool solve_sudoku_min_arity(int which_space);
-bool solve_sudoku_min_arity_cache(int which_space);
-bool solve_sudoku_dancing_links(int unused);
-bool solved();
-#endif
