--- Lab1/final/README
@@ -1,3 +1,0 @@
-A backtracking sudoku solver, four algorithms are shown.
-
-test cases are from http://people.csse.uwa.edu.au/gordon/sudokumin.php
