--- Lab1/final/test4
@@ -1,4 +1,0 @@
-000000010400000000020000000000050407008000300001090000300400200050100000000806000
-000000010400000000020000000000050604008000300001090000300400200050100000000807000
-000000012000035000000600070700000300000400800100000000000120000080000040050000600
-000000012003600000000007000410020000000500300700000600280000040000300500000000000
