--- Lab1/src/Sudoku/test1
@@ -1,1 +1,0 @@
-000000010400000000020000000000050407008000300001090000300400200050100000000806000
