--- Lab1/src/Sudoku/test10
@@ -1,10 +1,0 @@
-000000010400000000020000000000050407008000300001090000300400200050100000000806000
-000000010400000000020000000000050604008000300001090000300400200050100000000807000
-000000012000035000000600070700000300000400800100000000000120000080000040050000600
-000000012003600000000007000410020000000500300700000600280000040000300500000000000
-000000012008030000000000040120500000000004700060000000507000300000620000000100000
-000000012040050000000009000070600400000100000000000050000087500601000300200000000
-000000012050400000000000030700600400001000000000080000920000800000510700000003000
-000000012300000060000040000900000500000001070020000000000350400001400800060000000
-000000012400090000000000050070200000600000400000108000018000000000030700502000000
-000000012500008000000700000600120000700000450000030000030000800000500700020000000
